refactor: article 관련 수정
DB에 게시글을 올리기 위해 타입, 필드 수정

+++ src/main/java/codesquad/codestagram/controller/ArticleController.java
@@ -31,7 +31,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{id}")
-    public String show(@PathVariable int id, Model model) {
+    public String show(@PathVariable Long id, Model model) {
         Article article = articleService.findOneArticle(id).get();
         model.addAttribute("article", article);
         return "/qna/show";
